storage: move store bootstrap to own file
Release note: None
